requirements.txt
# BasuniAI
python-dotenv>=1.0.0
PyYAML>=6.0
discord.py>=2.3.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
openai>=1.0.0
